// src/story/dto/create-story-image.dto.ts
// dto for facebook story image
export class CreateStoryImageDto {
    _id: string;
    image: string;
    image_scale: string;
    image_position_x: string;
    image_position_y: string;
    post: string;
    title: string;
    text_color: string;
    background_color: string;
    text_size: string;
    text_position: string;
    text_padding: string;
    text_family: string;
    text_style: string;
    text_weight: string;
    text_decoration: string;
    text_shadow: string;
    text_outline: string;
    text_align: string;
    created_by: string;
    updated_by: string;
    CreatedAt: Date;
    UpdatedAt: Date;
    __v: number;
}